More doubles

+++ Source/FieldSolver/FiniteDifferenceSolver/EvolveE.cpp
@@ -432,7 +432,7 @@ void FiniteDifferenceSolver::EvolveECylindrical (
             amrex::Gpu::synchronize();
             wt = amrex::second() - wt;
             amrex::HostDevice::Atomic::Add( &(*cost)[mfi.index()], wt);
-            amrex::HostDevice::Atomic::Add( &(*(warpx.g_temp)[lev])[mfi.index()], wt);
+            amrex::HostDevice::Atomic::Add( &(*(warpx.g_temp)[lev])[mfi.index()], double(wt));
         }
     } // end of loop over grid/tiles
